firmware/server/legacy/requirements.txt
paho-mqtt>=2.0
Pillow>=10.0
